style: reduce the number of files in root directory (because you need to scroll too many times before reading the readme)
