+++ .github/dependabot.yml
@@ -4,10 +4,12 @@ updates:
     directory: /
     schedule:
       interval: monthly
+
   - package-ecosystem: gomod
     directory: /
     schedule:
       interval: weekly
+
   - package-ecosystem: npm
     directory: /docs
     schedule:
