+++ .github/workflows/doc.yml
@@ -1,4 +1,4 @@
-name: Docs helpers validation
+name: Documentation
 
 on:
   pull_request:
@@ -9,7 +9,7 @@ on:
       - 'docs/**'
 
 jobs:
-  validate-docs-helpers:
+  build-doc:
     runs-on: ubuntu-latest
     defaults:
       run:
@@ -28,20 +28,37 @@ jobs:
       - name: Build Docusaurus project
         run: npm run build
 
-      # - name: Run docs helpers checks
-      #   run: |
-      #     node docs/scripts/check-duplicates-in-category.js
-      #     node docs/scripts/check-similar-exists.js
-      #     node docs/scripts/check-filename-matches-frontmatter.js
-      #     node docs/scripts/check-similar-keys-exist-in-directory.js
-      #     node docs/scripts/check-cross-references.js
 
-      # - name: Sync function signatures
-      #   run: npm run sync:function-signatures
-      # - name: Commit and push changes
-      #   uses: stefanzweifel/git-auto-commit-action@v5
-      #   with:
-      #     commit_message: 'docs: sync function signatures from source'
-      #     commit_user_name: github-actions[bot]
-      #     commit_user_email: github-actions[bot]@users.noreply.github.com
-      #     file_pattern: docs/data/*.md
+  # validate-doc:
+  #   runs-on: ubuntu-latest
+  #   defaults:
+  #     run:
+  #       working-directory: ./docs
+
+  #   steps:
+  #     - name: Checkout repository
+  #       uses: actions/checkout@v5
+
+  #     - name: Setup Node.js
+  #       uses: actions/setup-node@v6
+
+  #     - name: Install dependencies
+  #       run: npm i
+
+  #     - name: Run docs helpers checks
+  #       run: |
+  #         node docs/scripts/check-duplicates-in-category.js
+  #         node docs/scripts/check-similar-exists.js
+  #         node docs/scripts/check-filename-matches-frontmatter.js
+  #         node docs/scripts/check-similar-keys-exist-in-directory.js
+  #         node docs/scripts/check-cross-references.js
+
+  #     - name: Sync function signatures
+  #       run: npm run sync:function-signatures
+  #     - name: Commit and push changes
+  #       uses: stefanzweifel/git-auto-commit-action@v5
+  #       with:
+  #         commit_message: 'docs: sync function signatures from source'
+  #         commit_user_name: github-actions[bot]
+  #         commit_user_email: github-actions[bot]@users.noreply.github.com
+  #         file_pattern: docs/data/*.md
