+++ benchmark/seq_benchmark_test.go
@@ -1,6 +1,6 @@
 //go:build go1.23
 
-package it
+package benchmark
 
 import (
 	"fmt"
@@ -8,25 +8,27 @@ import (
 	"math/rand/v2"
 	"strconv"
 	"testing"
+
+	"github.com/samber/lo/it"
 )
 
-var lengths = []int{10, 100, 1000}
+var itLengths = []int{10, 100, 1000}
 
-func BenchmarkChunk(b *testing.B) {
-	for _, n := range lengths {
+func BenchmarkItChunk(b *testing.B) {
+	for _, n := range itLengths {
 		strs := genStrings(n)
 		b.Run(fmt.Sprintf("strings_%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = Chunk(strs, 5)
+				_ = it.Chunk(strs, 5)
 			}
 		})
 	}
 
-	for _, n := range lengths {
+	for _, n := range itLengths {
 		ints := genInts(n)
 		b.Run(fmt.Sprintf("ints%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = Chunk(ints, 5)
+				_ = it.Chunk(ints, 5)
 			}
 		})
 	}
@@ -52,99 +54,100 @@ func genInts(n int) iter.Seq[int] {
 	}
 }
 
-func BenchmarkFlatten(b *testing.B) {
-	for _, n := range lengths {
+func BenchmarkItFlatten(b *testing.B) {
+	for _, n := range itLengths {
 		ints := make([]iter.Seq[int], 0, n)
 		for i := 0; i < n; i++ {
 			ints = append(ints, genInts(n))
 		}
 		b.Run(fmt.Sprintf("ints_%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = Flatten(ints)
+				_ = it.Flatten(ints)
 			}
 		})
 	}
 
-	for _, n := range lengths {
+	for _, n := range itLengths {
 		strs := make([]iter.Seq[string], 0, n)
 		for i := 0; i < n; i++ {
 			strs = append(strs, genStrings(n))
 		}
 		b.Run(fmt.Sprintf("strings_%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = Flatten(strs)
+				_ = it.Flatten(strs)
 			}
 		})
 	}
 }
 
-func BenchmarkDrop(b *testing.B) {
-	for _, n := range lengths {
+func BenchmarkItDrop(b *testing.B) {
+	for _, n := range itLengths {
 		strs := genStrings(n)
 		b.Run(fmt.Sprintf("strings_%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = Drop(strs, n/4)
+				_ = it.Drop(strs, n/4)
 			}
 		})
 	}
 
-	for _, n := range lengths {
+	for _, n := range itLengths {
 		ints := genInts(n)
 		b.Run(fmt.Sprintf("ints%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = Drop(ints, n/4)
+				_ = it.Drop(ints, n/4)
 			}
 		})
 	}
 }
 
-func BenchmarkDropWhile(b *testing.B) {
-	for _, n := range lengths {
+func BenchmarkItDropWhile(b *testing.B) {
+	for _, n := range itLengths {
 		strs := genStrings(n)
 		b.Run(fmt.Sprintf("strings_%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = DropWhile(strs, func(v string) bool { return len(v) < 4 })
+				_ = it.DropWhile(strs, func(v string) bool { return len(v) < 4 })
 			}
 		})
 	}
 
-	for _, n := range lengths {
+	for _, n := range itLengths {
 		ints := genInts(n)
 		b.Run(fmt.Sprintf("ints%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = DropWhile(ints, func(v int) bool { return i < 10_000 })
+				_ = it.DropWhile(ints, func(v int) bool { return i < 10_000 })
 			}
 		})
 	}
 }
 
-func BenchmarkDropByIndex(b *testing.B) {
-	for _, n := range lengths {
+func BenchmarkItDropByIndex(b *testing.B) {
+	for _, n := range itLengths {
 		strs := genStrings(n)
 		b.Run(fmt.Sprintf("strings_%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = DropByIndex(strs, n/4)
+				_ = it.DropByIndex(strs, n/4)
 			}
 		})
 	}
 
-	for _, n := range lengths {
+	for _, n := range itLengths {
 		ints := genInts(n)
 		b.Run(fmt.Sprintf("ints%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = DropByIndex(ints, n/4)
+				_ = it.DropByIndex(ints, n/4)
 			}
 		})
 	}
 }
 
-func BenchmarkReplace(b *testing.B) {
+func BenchmarkItReplace(b *testing.B) {
 	lengths := []int{1_000, 10_000, 100_000}
+
 	for _, n := range lengths {
 		strs := genStrings(n)
 		b.Run(fmt.Sprintf("strings_%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = Replace(strs, "321321", "123123", 10)
+				_ = it.Replace(strs, "321321", "123123", 10)
 			}
 		})
 	}
@@ -153,7 +156,7 @@ func BenchmarkReplace(b *testing.B) {
 		ints := genInts(n)
 		b.Run(fmt.Sprintf("ints%d", n), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				_ = Replace(ints, 321321, 123123, 10)
+				_ = it.Replace(ints, 321321, 123123, 10)
 			}
 		})
 	}
